// src/generated/models/NewPrecodeProdutoModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface NewPrecodeProdutoBase {
  new_precodeprodutoid: string;
  new_aceitadesconto?: boolean;
  new_cadastro?: string;
  new_categoriadeprodutofx?: string;
  new_categoriadeprodutoid?: string;
  new_datadaultimaatualizacaodeprecosautomatica?: string;
  new_descontodekitefetivo?: number;
  new_descontodekitpercentual?: number;
  new_descontopercentualdecompra?: number;
  new_descricao?: string;
  new_guiacomercialurl?: string;
  new_kit?: boolean;
  new_mappedmodified?: string;
  new_markup?: number;
  new_name?: string;
  new_nomedofabricante?: string;
  new_nomedoproduto?: string;
  new_novadescricaoflag?: boolean;
  new_organizacaoproprietaria?: number;
  new_organizacaoproprietariakit?: number;
  new_parent?: string;
  new_precobase?: number;
  new_precocominstalacao?: number;
  new_precodecompra?: number;
  new_precodeservicodeinstalacao?: number;
  new_precodeservicodeinstalacaodecimalfx?: number;
  new_precodevenda?: number;
  new_referenciadoproduto?: string;
  new_requerinstalacao?: boolean;
  new_servicoadicionado?: boolean;
  new_taxadeservico?: number;
  new_taxafixadeservico?: number;
  new_tipodesistema?: number;
  new_validade?: string;
  "new_AtividadedeCopiadeProduto@odata.bind"?: string;
  "new_AtividadedeTrocadeItemdeKit@odata.bind"?: string;
  "new_AtividadeTrocadePalavras@odata.bind"?: string;
  "new_CotacaoTemporaria@odata.bind"?: string;
  "new_Fornecedor@odata.bind"?: string;
  "new_ModelodeProduto@odata.bind"?: string;
  "new_ProdutoDesigner@odata.bind"?: string;
  "new_Promocao@odata.bind"?: string;
  "new_ServicoparaadicionarDEV@odata.bind"?: string;
  statecode?: number;
  statuscode?: number;
}

export interface NewPrecodeProduto extends NewPrecodeProdutoBase {
  _new_atividadedecopiadeproduto_value?: string;
  _new_atividadedetrocadeitemdekit_value?: string;
  _new_atividadetrocadepalavras_value?: string;
  _new_cotacaotemporaria_value?: string;
  _new_fornecedor_value?: string;
  _new_modelodeproduto_value?: string;
  _new_produtodesigner_value?: string;
  _new_promocao_value?: string;
  _new_servicoparaadicionardev_value?: string;
}